+++ lib/vagrant-registration/action/register.rb
@@ -11,7 +11,10 @@ module VagrantPlugins
         end
 
         def call(env)
-          @app.call(env)
+          # Vbguest plugin (if present) is called next. Therefore registration
+          # needs to be done first. This does not work with the default
+          # 'action_register' hook.
+          @app.call(env) unless Plugin.vbguest_plugin?
 
           # Configuration from Vagrantfile
           config = env[:machine].config.registration
@@ -38,6 +41,8 @@ module VagrantPlugins
           end
 
           @logger.debug('Registration is skipped due to the configuration') if config.skip
+
+          @app.call(env) if Plugin.vbguest_plugin?
         end
 
         private